Moves game logic to game.rb
